requirements.txt
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
schedule==1.2.0
pyyaml==6.0.1
